% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{file.to.regnet}
\alias{file.to.regnet}
\title{This function reads in panda C output and converts to a reg.net mxn matrix}
\usage{
file.to.regnet(file.name)
}
\arguments{
\item{path}{a path to the folder containing the network files (and only the network files)}

\item{keyword}{a string to be used as a key for distinguishing group 1 from group 2}
}
\description{
This function reads in panda C output and converts to a reg.net mxn matrix
}
\examples{
example1
}
\keyword{keywords}

