% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{load.null.tms}
\alias{load.null.tms}
\title{This function generates n null transformation matrices from a folder of networks}
\usage{
load.null.tms(dir.path, keyword = "sub1")
}
\arguments{
\item{keyword}{a string to be used as a key for distinguishing group 1 from group 2}

\item{path}{a path to the folder containing the network files (and only the network files)}
}
\description{
This function generates n null transformation matrices from a folder of networks
}
\examples{
example1
}
\keyword{keywords}

