% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{ssodm}
\alias{ssodm}
\title{Sum of squared off-diagonal mass}
\usage{
ssodm(tm)
}
\arguments{
\item{tm}{a transition matrix for two bipartite networks}
}
\description{
This function calculates the off-diagonal sum of squared mass for a transition matrix
}
\examples{
data(yeast.panda)
t.matrix <- transformation.matrix(yeast.panda$cell.cycle, yeast.panda$stress.response)
ssodm(t.matrix)
}
\keyword{keywords}

