% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_functions.R
\name{hcl.heatmap.plot}
\alias{hcl.heatmap.plot}
\title{Transformation matrix plot}
\usage{
hcl.heatmap.plot(x, method = "pearson")
}
\arguments{
\item{method}{distance metric for hierarchical clustering.  Default is "Pearson correlation"}

\item{net1}{starting network, a genes by transcription factors data.frame with scores for confidence in the existence of edges between}
}
\description{
This function plots a hierachically clustered heatmap and corresponding dendrogram of a transaction matrix
}
\examples{
data(yeast.panda)
t.matrix <- transformation.matrix(yeast.panda$cell.cycle, yeast.panda$stress.response)
hcl.heatmap.plot(t.matrix, method="pearson")
}
\keyword{keywords}

